# Conflicts:
#	app/build.gradle.kts
#	app/src/main/java/org/merakilearn/FileChooserActivity.kt
#	app/src/main/java/org/merakilearn/InstallReferrerManager.kt
#	app/src/main/java/org/merakilearn/MainActivity.kt
#	app/src/main/java/org/merakilearn/datasource/ClassesRepo.kt
#	app/src/main/java/org/merakilearn/datasource/PlaygroundRepo.kt
#	app/src/main/java/org/merakilearn/datasource/network/SaralApi.kt
#	app/src/main/java/org/merakilearn/datasource/network/UserUpdateName.kt
#	app/src/main/java/org/merakilearn/datasource/network/model/UploadCredentials.kt
#	app/src/main/java/org/merakilearn/ui/ScratchActivity.kt
#	app/src/main/java/org/merakilearn/ui/onboarding/SelectCourseFragment.kt
#	app/src/main/java/org/merakilearn/ui/playground/PlaygroundFragment.kt
#	app/src/main/java/org/merakilearn/ui/playground/PlaygroundViewModel.kt
#	app/src/main/res/values/array.xml
#	app/src/main/res/values/strings.xml
#	buildSrc/src/main/kotlin/Dependencies.kt
#	commonUI/src/main/java/org/navgurukul/commonui/platform/SvgLoader.kt
#	core/src/main/res/xml/remote_config_defaults.xml
#	learn/src/main/java/org/navgurukul/learn/courses/db/CoursesDatabase.kt
#	learn/src/main/java/org/navgurukul/learn/courses/db/models/BaseCourseContent.kt
#	learn/src/main/java/org/navgurukul/learn/courses/repository/LearnRepo.kt
#	learn/src/main/java/org/navgurukul/learn/di/LearnModules.kt
#	learn/src/main/java/org/navgurukul/learn/editor/LineNumberEditText.kt
#	learn/src/main/java/org/navgurukul/learn/ui/learn/AssessmentFragment.kt
#	learn/src/main/java/org/navgurukul/learn/ui/learn/ExerciseFragment.kt
#	learn/src/main/java/org/navgurukul/learn/ui/learn/LearnFragment.kt
#	learn/src/main/java/org/navgurukul/learn/ui/learn/adapter/ExerciseContentAdapter.kt
#	learn/src/main/java/org/navgurukul/learn/ui/learn/viewholder/CodeCourseViewHolder.kt
#	learn/src/main/java/org/navgurukul/learn/ui/learn/viewholder/QuestionCodeCourseViewHolder.kt
#	learn/src/main/res/drawable/ic_baseline_arrow_right_24.xml
#	learn/src/main/res/drawable/ic_baseline_restart_alt_24.xml
#	learn/src/main/res/layout/fragment_learn.xml
#	learn/src/main/res/values/colors.xml
#	python/src/main/java/org/navgurukul/playground/editor/PythonEditorActivity.kt

+++ learn/src/main/java/org/navgurukul/learn/ui/learn/viewholder/AssessmentFragmentViewModel.kt
@@ -314,7 +314,7 @@ class AssessmentFragmentViewModel (
             return clickedOption.id ==
                     (allAssessmentContentList
                         .find { it.component == BaseCourseContent.COMPONENT_SOLUTION } as SolutionBaseCourseContent)
-                        .value
+                        .value.toInt()
 
         }catch (e: Exception){
             return false
